+++ lib/src/paths/lib.dart
@@ -1,0 +1,5 @@
+export 'my_file_root.dart';
+export 'my_user_data_directory_store.dart';
+export 'my_user_data_directory_validator.dart';
+export 'my_user_data_files_migrator.dart';
+export 'my_user_data_paths.dart';